backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Tanner Schmoekel"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
asyncpg = "^0.27.0"
bcrypt = "^4.0.1"
fastapi = "^0.115.5"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.5"
dnspython = "^2.2.1"
motor = "^3.1.1"
requests = "^2.28.2"
python-dateutil = "^2.8.2"
sentry-sdk = {extras = ["fastapi"], version = "^1.15.0"}
tortoise-orm = "^0.19.3"
jinja2 = "^3.1.2"
pytz = "^2022.7.1"
twilio = "^7.17.0"
phonenumbers = "^8.13.7"
python-decouple = "^3.8"
python-dotenv = "^1.0.0"
sendgrid = "^6.10.0"
boto3 = "^1.35.63"
python-slugify = "^8.0.1"
mailersend = "^0.5.1"
fastapi-utils = "^0.2.1"
loguru = "^0.7.1"
redis="<5.0.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
psycopg2-binary = "^2.9.9"
fastapi-cache2 = {extras = ["redis"], version = "0.2.2"}
stripe="^9.12.0"
aerich = "^0.7.1"
httpx = "^0.24.0"
cloudinary = "1.42.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.2.0"
black = "^23.3.0"
flake8 = "4.0.1"
pytest = "^7.3.1"
pytest-flake8 = "^1.1.1"
pytest-cov = "^4.0.0"
trio = "^0.22.0"
pre-commit = "^2.21.0"
pytest-mock = "^3.10.0"
factory-boy = "^3.2.1"
faker = "^18.6.2"
async-factory-boy = "^1.0.0"
aws-cdk-lib = "^2.92.0"

[tool.aerich]
tortoise_orm = "src.database.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
skip-string-normalization = true
force-exclude = '''
  (migrations\/.*\.py$|manage\.py|(w|a)sgi\.py|apps\.py|urls\.py|settings\.py)
'''

[tool.isort]
atomic = true
case_sensitive = true
extend_skip = [".json", ".md"]
skip = ["manage.py", "wsgi.py", "asgi.py", "apps.py", "urls.py"]
skip_glob = ["*/migrations/*"]
filter_files = true
import_heading_firstparty = "Internal imports"
import_heading_stdlib = "Python imports"
import_heading_thirdparty = "Pip imports"
known_startup = ["startup"]
line_length = 120
lines_after_imports = 2
profile = "black"
py_version = "auto"
remove_redundant_aliases = true
sections = [
    "FUTURE",
    "STARTUP",
    "STDLIB",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]
